# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample for ptushki-backend API documentation",
    "version": "1.0.0",
    "title": "Ptushki API"
  },
...
  "schemes": ["http"],
  "paths": {
...
    "/observations/export/{type}": {
      "post": {
        "tags": ["Observation controller"],
        "summary": "Generate export of specified type",
        "produces": ["application/json", "application/pdf", "application/vnd.ms-excel"],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "type of export, planned to support xls, pdf, euring",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/observations/import/{type}": {
      "post": {
        "tags": ["Observation controller"],
        "summary": "Import data to observations from extranal data of specified type",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "description": "type of import, planned to support xls, euring",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
    "/observations/{id}": {
      "get": {
        "tags": ["Observation controller"],
        "produces": ["application/json"],
        "summary": "Get observation by id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of observation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Observation"
            }
          }
        }
      },
      "put": {
        "tags": ["Observation controller"],
        "summary": "Edit observation with provided id",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of observation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Observation"
            }
          }
        }
      },
      "delete": {
        "tags": ["Observation controller"],
        "summary": "Remove observation by provided id",
        "produces": ["application/json"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of observation",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Observation"
            }
          }
        }
      }
    },
    "/observations/{id}/export/{type}": {
      "get": {
        "tags": ["Observation controller"],
        "produces": ["application/json", "application/pdf"],
        "summary": "Generate specified type export for current entity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of observation",
            "required": true,
            "type": "string"
          },
          {
            "name": "type",
            "in": "path",
            "description": "type of export, planned to support pdf, euring",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "AccuracyOfCoordinates": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "AccuracyOfDate": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "AccuracyOfPullusAge": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "Age": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "mentionedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "concludedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "mentionedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "concludedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "basaRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasaRing"
          }
        }
      }
    },
    "BroodSize": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "CatchingLures": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "CatchingMethod": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "Circumstances": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "CircumstancesPresumed": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "Condition": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "EURINGCodeIdentifier": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "OtherMarksInformation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "PrimaryIdentificationMethod": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "Manipulated": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "MetalRingInformation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "MovedBeforeTheCapture": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "PlaceCode": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "PullusAge": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "RingingScheme": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "country": {
          "type": "string"
        },
        "center": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
    "Sex": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "mentionedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "concludedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "mentionedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "concludedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "basaRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasaRing"
          }
        }
      }
    },
    "Species": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "letterCode": {
          "type": "string"
        },
        "species": {
          "type": "string"
        },
        "ordo": {
          "type": "string"
        },
        "family": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "mentionedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "concludedInRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "mentionedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "concludedInObservation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "basaRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasaRing"
          }
        }
      }
    },
    "Status": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        }
      }
    },
    "StatusOfRing": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        }
      }
    },
    "VerificationOfTheMetalRing": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "desc_eng": {
          "type": "string"
        },
        "desc_rus": {
          "type": "string"
        },
        "desc_byn": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        }
      }
    },
...
    "BasaRing": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "ringNumber": {
          "type": "string"
        },
        "species": {
          "$ref": "#/definitions/Species"
        },
        "sex": {
          "$ref": "#/definitions/Sex"
        },
        "age": {
          "$ref": "#/definitions/Age"
        },
        "ringer": {
          "$ref": "#/definitions/User"
        },
        "region": {
          "type": "string"
        },
        "district": {
          "type": "string"
        },
        "place": {
          "type": "string"
        },
        "coordinates": {
          "type": "string"
        },
        "note": {
          "type": "string"
        }
      }
    },
    "Observation": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "ring": {
          "$ref": "#/definitions/Ring"
        },
        "ringMentioned": {
          "type": "string"
        },
        "finder": {
          "$ref": "#/definitions/User"
        },
        "speciesMentioned": {
          "$ref": "#/definitions/Species"
        },
        "speciesConcluded": {
          "$ref": "#/definitions/Species"
        },
        "sexMentioned": {
          "$ref": "#/definitions/Sex"
        },
        "sexConcluded": {
          "$ref": "#/definitions/Sex"
        },
        "ageMentioned": {
          "$ref": "#/definitions/Age"
        },
        "ageConcluded": {
          "$ref": "#/definitions/Age"
        },
        "manipulated": {
          "$ref": "#/definitions/Manipulated"
        },
        "movedBeforeTheCapture": {
          "$ref": "#/definitions/MovedBeforeTheCapture"
        },
        "catchingMethod": {
          "$ref": "#/definitions/CatchingMethod"
        },
        "catchingLures": {
          "$ref": "#/definitions/CatchingLures"
        },
        "status": {
          "$ref": "#/definitions/Status"
        },
        "pullusAge": {
          "$ref": "#/definitions/PullusAge"
        },
        "accuracyOfPullusAge": {
          "$ref": "#/definitions/AccuracyOfPullusAge"
        },
        "geographicalCoordinates": {
          "type": "string"
        },
        "accuracyOfCoordinates": {
          "$ref": "#/definitions/AccuracyOfCoordinates"
        },
        "condition": {
          "$ref": "#/definitions/Condition"
        },
        "circumstances": {
          "$ref": "#/definitions/Circumstances"
        },
        "circumstancesPresumed": {
          "$ref": "#/definitions/CircumstancesPresumed"
        },
        "date": {
          "$ref": "#/definitions/Date"
        },
        "accuracyOfDate": {
          "$ref": "#/definitions/AccuracyOfDate"
        },
        "distance": {
          "type": "string"
        },
        "direction": {
          "type": "string"
        },
        "elapsedTime": {
          "type": "string"
        },
        "colorRing": {
          "type": "string"
        },
        "placeName": {
          "type": "string"
        },
        "remarks": {
          "type": "string"
        },
        "verified": {
          "type": "boolean"
        }
      }
    },
    "Ring": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "identificationNumber": {
          "type": "string"
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "ringingScheme": {
          "$ref": "#/definitions/RingingScheme"
        },
        "primaryIdentificationMethod": {
          "$ref": "#/definitions/PrimaryIdentificationMethod"
        },
        "verificationOfTheMetalRing": {
          "$ref": "#/definitions/VerificationOfTheMetalRing"
        },
        "metalRingInformation": {
          "$ref": "#/definitions/MetalRingInformation"
        },
        "otherMarksInformation": {
          "$ref": "#/definitions/OtherMarksInformation"
        },
        "speciesMentioned": {
          "$ref": "#/definitions/Species"
        },
        "speciesConcluded": {
          "$ref": "#/definitions/Species"
        },
        "manipulated": {
          "$ref": "#/definitions/Manipulated"
        },
        "movedBeforeTheCapture": {
          "$ref": "#/definitions/MovedBeforeTheCapture"
        },
        "catchingMethod": {
          "$ref": "#/definitions/CatchingMethod"
        },
        "catchingLures": {
          "$ref": "#/definitions/CatchingLures"
        },
        "sexMentioned": {
          "$ref": "#/definitions/Sex"
        },
        "sexConcluded": {
          "$ref": "#/definitions/Sex"
        },
        "ageMentioned": {
          "$ref": "#/definitions/Age"
        },
        "ageConcluded": {
          "$ref": "#/definitions/Age"
        },
        "status": {
          "$ref": "#/definitions/Status"
        },
        "broodSize": {
          "$ref": "#/definitions/BroodSize"
        },
        "pullusAge": {
          "$ref": "#/definitions/PullusAge"
        },
        "accuracyOfPullusAge": {
          "$ref": "#/definitions/AccuracyOfPullusAge"
        },
        "geographicalCoordinates": {
          "type": "string"
        },
        "placeCode": {
          "$ref": "#/definitions/PlaceCode"
        },
        "accuracyOfCoordinates": {
          "$ref": "#/definitions/AccuracyOfCoordinates"
        },
        "condition": {
          "$ref": "#/definitions/Condition"
        },
        "circumstances": {
          "$ref": "#/definitions/Circumstances"
        },
        "circumstancesPresumed": {
          "$ref": "#/definitions/CircumstancesPresumed"
        },
        "date": {
          "$ref": "#/definitions/Date"
        },
        "accuracyOfDate": {
          "$ref": "#/definitions/AccuracyOfDate"
        },
        "euringCodeIdentifier": {
          "$ref": "#/definitions/EURINGCodeIdentifier"
        },
        "remarks": {
          "type": "string"
        },
        "ringerInformation": {
          "$ref": "#/definitions/User"
        },
        "statusOfRing": {
          "$ref": "#/definitions/StatusOfRing"
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "hash": {
          "type": "string"
        },
        "salt": {
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/UserRole"
        },
        "firstName": {
          "type": "string"
        },
        "lastName": {
          "type": "string"
        },
        "ring": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Ring"
          }
        },
        "observation": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Observation"
          }
        },
        "basaRing": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/BasaRing"
          }
        }
      }
    },
    "UserRole": {
      "type": "array",
      "items": {
        "type": "string",
        "enum": ["observer", "ringer", "scientist", "moderator", "admin"]
      }
    },
    "Date": {
      "type": "object",
      "description": "https://developer.mozilla.org/ru/docs/Web/JavaScript/Reference/Global_Objects/Date"
    }
  }
}